requirements.txt
python-dotenv==1.0.0
web3==7.4.0
